fix: 解决Agent心跳401认证问题，完善认证白名单配置
问题原因：
- SecurityConfig和InterceptorConfig需要同时配置白名单
- 只配置一处导致仍然返回401/403

修改内容：
1. InterceptorConfig.java 添加heartbeat/agents/registry/monitor路径白名单
2. 更新开发规范文档，增加认证与权限规范章节

涉及文件：
- backend/src/main/java/com/aipal/config/InterceptorConfig.java
- docs/development-standard.md

+++ backend/src/main/java/com/aipal/service/HeartbeatManagementServiceImpl.java
@@ -48,7 +48,7 @@ public class HeartbeatManagementServiceImpl implements HeartbeatManagementServic
 
         // 获取Agent的heartbeatTimeout配置
         AgentRegistration registration = getRegistration(agentCode, instanceId);
-        int timeoutSeconds = DEFAULT_HEARTBEAT_TIMEOUT.getSeconds();
+        int timeoutSeconds = (int) DEFAULT_HEARTBEAT_TIMEOUT.getSeconds();
         if (registration != null && registration.getHeartbeatTimeout() != null) {
             timeoutSeconds = registration.getHeartbeatTimeout();
         }
@@ -132,7 +132,7 @@ public class HeartbeatManagementServiceImpl implements HeartbeatManagementServic
                     String instanceId = parts.length > 1 ? parts[1] : "default";
 
                     AgentRegistration registration = getRegistration(agentCode, instanceId);
-                    int timeoutSeconds = DEFAULT_HEARTBEAT_TIMEOUT.getSeconds();
+                    int timeoutSeconds = (int) DEFAULT_HEARTBEAT_TIMEOUT.getSeconds();
                     if (registration != null && registration.getHeartbeatTimeout() != null) {
                         timeoutSeconds = registration.getHeartbeatTimeout();
                     }
@@ -165,7 +165,7 @@ public class HeartbeatManagementServiceImpl implements HeartbeatManagementServic
 
             // 获取该Agent的超时配置
             AgentRegistration registration = getRegistration(agentCode, instanceId);
-            int timeoutSeconds = DEFAULT_HEARTBEAT_TIMEOUT.getSeconds();
+            int timeoutSeconds = (int) DEFAULT_HEARTBEAT_TIMEOUT.getSeconds();
             if (registration != null && registration.getHeartbeatTimeout() != null) {
                 timeoutSeconds = registration.getHeartbeatTimeout();
             }
